fix: incorrect downloaded size at end of downloads
The downloaded size reported at the end of all downloads was incorrect
on some configurations (presumably: not high-end CPUs).
It is due to the fact that lock() was used to update the UI each time
a file size was increased, but lock() does not guarantee FIFO!

We now use ConcurrentQueue<T> instead, which is thread-safe.

Fixes #65